backend/requirements/test.txt
# test.txt
-r base.txt
-r dev.txt

factory-boy==3.3.3
faker==40.11.1
iniconfig==2.3.0
pluggy==1.6.0
pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
